pyproject.toml
[tool.poetry]
name = "discord-dingomata"
version = "2.0.0"
description = "A discord bot that does a bunch of random things."
authors = ["SnazzyFox <12706268+snazzyfox@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "~3.11"
pytz = "^2023.3"
prettytable = "^3.8.0"
parsedatetime = "^2.6"
Unidecode = "^1.3.6"
tortoise-orm = {extras = ["accel", "asyncpg"], version = "^0.19.3"}
openai = "^0.27.8"
pydantic-settings = "^2.0.2"
async-lru = "^2.0.4"
hikari-lightbulb = "^2.3.3"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.mypy]
plugins = ["pydantic.mypy"]